pyproject.toml
[project]
name = "fastapi_app"
version = "0.1.0"
description = ""
authors = [
    {name = "piyush"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "psycopg[pool] (>=3.2.9,<4.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
uvicorn = "^0.35.0"
fastapi = "^0.116.1"

